Added Chart Full

+++ backEnd/src/services/categoryService.ts
@@ -16,25 +16,27 @@ export const getcategoryServices = async (_req: any, res: Response) => {
 
 export const createcategoryServices = async (req: any, res: Response) => {
   try {
+    let requestedUser = await User.findById(req.decoded.id)
+    if (!requestedUser) {
+      return res.status(401).send("Cannot find the user")
+    }
+    if (requestedUser.type != "Admin") {
+      return res.status(403).send("Unauthorized")
+    }
     let profile = req.body.profileImage;
     if (req.file) {
       profile = req.file.path.replace('\\', '/');
     }
-
-    console.log(req.body)
     const categoryData = {
       profile: profile,
       name: req.body.name,
     }
-
-    console.log(categoryData)
     const category = new categorys(categoryData);
     const result = await category.save();
-    res.status(201).json({ message: "Created Successfully", data: result })
+    return res.status(201).json({ message: "Created Successfully", data: result })
   } catch (err) {
-    res.send("An Error occured in create category");
-    console.log(err)
     logger.categoryInfoLogger.log('info', 'Error Create Category')
+    return res.send("An Error occured in create category");
   }
 };
 
@@ -51,6 +53,13 @@ export const findcategoryServices = async (req: any, res: Response) => {
 
 export const updatecategoryServices = async (req: any, res: Response) => {
   try {
+    let requestedUser = await User.findById(req.decoded.id)
+    if (!requestedUser) {
+      return res.status(401).send("Cannot find the user")
+    }
+    if (requestedUser.type != "Admin") {
+      return res.status(403).send("Unauthorized")
+    }
     const Category = await categorys.findById(req.params.id);
     if (!Category) {
       const error = new Error("Not Found");
@@ -70,22 +79,29 @@ export const updatecategoryServices = async (req: any, res: Response) => {
     category.name = req.body.name;
     category.price = req.body.price;
     const result = await category.save();
-    res.json({ message: "Updated Successfully!", data: result })
+    return res.json({ message: "Updated Successfully!", data: result })
   } catch (err) {
-    res.send("an error occured in Edit Category");
     console.log(err)
     logger.categoryErrorLogger.log('info', 'Error Update Category')
+    return res.send("an error occured in Edit Category");
   }
 };
 
 export const deletecategoryServices = async (req: any, res: Response) => {
   try {
+    let requestedUser = await User.findById(req.decoded.id)
+    if (!requestedUser) {
+      return res.status(401).send("Cannot find the user")
+    }
+    if (requestedUser.type != "Admin") {
+      return res.status(403).send("Unauthorized")
+    }
     await categorys.findById(req.params.id);
     await categorys.findByIdAndRemove(req.params.id);
-    res.json({ message: "category with id " + req.params.id + " removed." })
+    return res.json({ message: "category with id " + req.params.id + " removed." })
   } catch (err) {
-    res.send("An Error Occured During Delete Category")
     console.log(err)
     logger.categoryErrorLogger.log('info', 'Error Delete Category')
+    return res.send("An Error Occured During Delete Category")
   }
 };
